+++ cg/services/illumina_services/post_processing/validation.py
@@ -4,7 +4,7 @@ from pathlib import Path
 from cg.constants.demultiplexing import DemultiplexingDirsAndFiles
 from cg.exc import FlowCellError, MissingFilesError
 from cg.models.run_devices.illumina_run_directory_data import IlluminaRunDirectoryData
-from cg.services.illumina_services.illumina_post_processing_service.utils import (
+from cg.services.illumina_services.post_processing.utils import (
     get_sample_fastqs_from_flow_cell,
 )
