+++ tests/services/illumina_services/illumina_post_processing_service/conftest.py
@@ -7,7 +7,7 @@ import pytest
 from cg.apps.housekeeper.hk import HousekeeperAPI
 from cg.constants.devices import DeviceType
 from cg.models.run_devices.illumina_run_directory_data import IlluminaRunDirectoryData
-from cg.services.illumina_services.illumina_post_processing_service.illumina_post_processing_service import (
+from cg.services.illumina_services.post_processing.post_processing_service import (
     IlluminaPostProcessingService,
 )
 from cg.store.models import IlluminaFlowCell
