+++ tests/services/illumina_services/illumina_post_processing_service/test_illumina_post_processing_service.py
@@ -1,7 +1,7 @@
 """Module to test the illumina post processing service."""
 
 from cg.models.run_devices.illumina_run_directory_data import IlluminaRunDirectoryData
-from cg.services.illumina_services.illumina_post_processing_service.illumina_post_processing_service import (
+from cg.services.illumina_services.post_processing.post_processing_service import (
     IlluminaPostProcessingService,
 )
 from cg.store.models import (
